Divided ViewSet into APIViews / views.py file into (auth|post).py
