+++ catlas/api/permissions.py
@@ -1,0 +1,5 @@
+from rest_framework import permissions
+
+class IsOwner(permissions.BasePermission):
+    def has_object_permission(self, request, view, obj):
+        return request.data['user'] == obj.user.id
