+++ catlas/api/urls.py
@@ -2,19 +2,19 @@ from django.urls import path, include
 from knox.views import LogoutView
 from rest_framework import routers
 
-from api import views
-
-router = routers.DefaultRouter()
-
-router.register(r'post', views.PostViewSet)
+from api.views import auth, post
 
 urlpatterns = [
     # Authentication, refactor into ViewSet
-    path('auth/login/', views.LoginAPI.as_view()),
-    path('auth/register/', views.RegistrationAPI.as_view()),
-    path('auth/reset/', views.ResetAPI.as_view()),
-    path('auth/check/', views.CheckAPI.as_view()),
-    path('auth/logout/', LogoutView.as_view())
-]
+    path('auth/login/', auth.LoginAPI.as_view()),
+    path('auth/register/', auth.RegistrationAPI.as_view()),
+    path('auth/reset/', auth.ResetAPI.as_view()),
+    path('auth/check/', auth.CheckAPI.as_view()),
+    path('auth/logout/', LogoutView.as_view()),
 
-urlpatterns += router.urls
+    path('post/', post.ListAPI.as_view()),
+    path('post/write/', post.CreateAPI().as_view()),
+    path('post/<int:pk>/', post.GetAPI.as_view()),
+    path('post/<int:pk>/update/', post.UpdateAPI.as_view()),
+    path('post/<int:pk>/delete/', post.DeleteAPI.as_view())
+]
